feat: add `oxc_index::non_zero` module.
