+++ crates/oxc_index/src/lib.rs
@@ -136,12 +136,14 @@ extern crate alloc;
 use core::fmt::Debug;
 use core::hash::Hash;
 
-pub mod slice;
 pub mod indexing;
+pub mod non_zero;
+pub mod slice;
 pub mod vec;
 
-pub use slice::{IndexBox, IndexSlice};
 pub use indexing::{IdxRangeBounds, IdxSliceIndex};
+pub use non_zero::NonZeroIdx;
+pub use slice::{IndexBox, IndexSlice};
 pub use vec::IndexVec;
 
 #[macro_use]
@@ -183,61 +185,3 @@ pub trait Idx: Copy + 'static + Ord + Debug + Hash {
     fn index(self) -> usize;
 }
 
-/// Represents a wrapped value convertible to and from a `usize`.
-///
-/// Generally you implement this via the [`define_non_zero_index_type!`] macro, rather
-/// than manually implementing it.
-///
-/// # Overflow
-///
-/// `NonZeroIdx` impls are allowed to be smaller than `usize`, which means converting
-/// `usize` to an `NonZeroIdx` implementation might have to handle overflow.
-///
-/// The way overflow is handled is up to the implementation of `NonZeroIdx`, but it's
-/// generally panicking, unless it was turned off via the
-/// `DISABLE_MAX_INDEX_CHECK` option in [`define_non_zero_index_type!`]. If you need more
-/// subtle handling than this, then you're on your own (or, well, either handle
-/// it earlier, or pick a bigger index type).
-///
-/// # Zero checks
-///
-/// `NonZeroIdx` impls are all, well "non-zero"; Which means it is always possible
-/// for a `usize` value to be 0 and we can not produce a valid `NonZeroIdx` from it.
-///
-///
-/// The way this is handled similar to `overflow` is up to the implementation of `NonZeroIdx`,
-/// but it's generally panicking, unless it was turned off via the
-/// `DISABLE_NON_ZERO_CHECK` option in [`define_non_zero_index_type!`]. If you need more
-/// subtle handling than this, then you're on your own (or, well, either handle
-/// it earlier, or pick a bigger index type).
-///
-pub trait NonZeroIdx: Copy + 'static + Ord + Debug + Hash {
-    /// Construct an Index from a `usize`. This is equivalent to `From<usize>`.
-    ///
-    /// # Panics
-    ///
-    /// Note that this will panic if `idx` does not fit or is 0 (unless checking has
-    /// been disabled, as mentioned above). Also note that `NonZeroIdx` implementations
-    /// are free to define what "fit" means as they desire(as long as it isn't zero).
-    fn from_usize_unchecked(idx: usize) -> Self;
-
-    /// Get the underlying index. This is equivalent to `Into<usize>`
-    fn index(self) -> usize;
-}
-
-/// A macro equivalent to the stdlib's `vec![]`, but producing an `IndexVec`.
-#[macro_export]
-macro_rules! index_vec {
-    ($($tokens:tt)*) => {
-        $crate::IndexVec::from_vec(vec![$($tokens)*])
-    }
-}
-
-/// A macro similar to the stdlib's `vec![]`, but producing an
-/// `Box<IndexSlice<I, [T]>>` (That is, an `IndexBox<I, [T]>`).
-#[macro_export]
-macro_rules! index_box {
-    ($($tokens:tt)*) => {
-        $crate::IndexVec::from_vec(vec![$($tokens)*]).into_boxed_slice()
-    }
-}
