+++ crates/oxc_index/src/macros.rs
@@ -1,3 +1,20 @@
+/// A macro equivalent to the stdlib's `vec![]`, but producing an `IndexVec`.
+#[macro_export]
+macro_rules! index_vec {
+    ($($tokens:tt)*) => {
+        $crate::IndexVec::from_vec(vec![$($tokens)*])
+    }
+}
+
+/// A macro similar to the stdlib's `vec![]`, but producing an
+/// `Box<IndexSlice<I, [T]>>` (That is, an `IndexBox<I, [T]>`).
+#[macro_export]
+macro_rules! index_box {
+    ($($tokens:tt)*) => {
+        $crate::IndexVec::from_vec(vec![$($tokens)*]).into_boxed_slice()
+    }
+}
+
 /// Generate the boilerplate for a newtyped index struct, for use with
 /// `IndexVec`.
 ///
